Completed changes to error/warning handling + other bits
Added a new test file test_fail.hex with purposely inserted errors
for testing purposes. This could be expanded.
Completed the error and warning handling structure usage and class
functions to use them.
Fixed a potential issue whereby a record missing the ':' record mark
still got its first character of the line deleted, even when the record
mark wasn't present. This was only considered to be worthy of a warning since
the rest of the record could potentially be decoded if it is a valid
record in all other aspects.