Further modification to support MSVC.
Used the _MSC_FULL_VER predefined macro to determine use of MSVC as
compiler. This selects stdio.h instead of cstdio as a header and uses
sprintf_s instead of snprintf which didn't seem to be supported.
Code now compiles with GNU g++ under minGW and MSVC under Qt Creator.